requirements.txt
flask
pymongo
dnspython
requests
python-dotenv
email-validator
werkzeug
flask-wtf
gunicorn